+++ app/who_is_there/__init__.py
@@ -1,2 +1,2 @@
 from .who_is_there import astros
-from .who_is_there import average
+from .astronauts import average
