+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.black]
-line-length = 120
+line-length = 80
 target-version = ['py311']
 extend-exclude = '''
 venv
